chore(core): remove convert v1 screens and rename v2
