+++ mobile-app/app/screens/AppNavigator/screens/Portfolio/screens/ConvertConfirmationScreen.tsx
@@ -1,11 +1,9 @@
-import { ThemedIcon, ThemedScrollView, ThemedSectionTitle, ThemedText, ThemedView } from '@components/themed'
+import { ThemedScrollViewV2, ThemedViewV2 } from '@components/themed'
 import { NavigationProp, useNavigation } from '@react-navigation/native'
 import { StackScreenProps } from '@react-navigation/stack'
 import BigNumber from 'bignumber.js'
 import { Dispatch, useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
-import { SubmitButtonGroup } from '@components/SubmitButtonGroup'
-import { SummaryTitle } from '@components/SummaryTitle'
 import { RootState } from '@store'
 import { hasTxQueued as hasBroadcastQueued } from '@store/ocean'
 import { hasTxQueued, transactionQueue } from '@store/transaction_queue'
@@ -13,15 +11,17 @@ import { tailwind } from '@tailwind'
 import { translate } from '@translations'
 import { PortfolioParamList } from '../PortfolioNavigator'
 import { ConversionMode } from './ConvertScreen'
-import { InfoRow, InfoType } from '@components/InfoRow'
-import { TextRow } from '@components/TextRow'
-import { TransactionResultsRow } from '@components/TransactionResultsRow'
-import { NumberRow } from '@components/NumberRow'
 import { NativeLoggingProps, useLogger } from '@shared-contexts/NativeLoggingProvider'
 import { onTransactionBroadcast } from '@api/transaction/transaction_commands'
 import { dfiConversionCrafter } from '@api/transaction/dfi_converter'
-import { WalletAddressRow } from '@components/WalletAddressRow'
 import { useAppDispatch } from '@hooks/useAppDispatch'
+import { SummaryTitleV2 } from '@components/SummaryTitleV2'
+import { SubmitButtonGroupV2 } from '@components/SubmitButtonGroupV2'
+import { View } from 'react-native'
+import { useWalletContext } from '@shared-contexts/WalletContext'
+import { useAddressLabel } from '@hooks/useAddressLabel'
+import { NumberRowV2 } from '@components/NumberRowV2'
+import { ConvertTokenUnit, getDisplayUnit } from '@screens/AppNavigator/screens/Portfolio/screens/ConvertScreen'
 
 type Props = StackScreenProps<PortfolioParamList, 'ConvertConfirmationScreen'>
 
@@ -35,6 +35,8 @@ export function ConvertConfirmationScreen ({ route }: Props): JSX.Element {
     amount,
     fee
   } = route.params
+  const { address } = useWalletContext()
+  const addressLabel = useAddressLabel(address)
   const hasPendingJob = useSelector((state: RootState) => hasTxQueued(state.transactionQueue))
   const hasPendingBroadcastJob = useSelector((state: RootState) => hasBroadcastQueued(state.ocean))
   const dispatch = useAppDispatch()
@@ -77,92 +79,113 @@ export function ConvertConfirmationScreen ({ route }: Props): JSX.Element {
   }
 
   return (
-    <ThemedScrollView style={tailwind('pb-4')}>
-      <ThemedView
-        dark={tailwind('bg-gray-800 border-b border-gray-700')}
-        light={tailwind('bg-white border-b border-gray-300')}
-        style={tailwind('flex-col px-4 py-8 mb-4')}
-      >
-        <SummaryTitle
+    <ThemedScrollViewV2 style={tailwind('pb-4')}>
+      <ThemedViewV2 style={tailwind('flex-col px-5 py-8')}>
+        <SummaryTitleV2
+          title={translate('screens/ConvertConfirmScreen', 'You are converting to {{unit}}', { unit: getDisplayUnit(targetUnit) })}
           amount={amount}
-          suffix={mode === 'utxosToAccount' ? 'DFI (UTXO)' : 'DFI (Token)'}
-          suffixType='component'
           testID='text_convert_amount'
-          title={translate('screens/ConvertConfirmScreen', 'You are converting')}
-        >
-          <ThemedText
-            light={tailwind('text-gray-500')}
-            dark={tailwind('text-gray-400')}
-            style={tailwind('text-sm')}
-            testID='convert_amount_source_suffix'
-          >
-            {sourceUnit}
-          </ThemedText>
-          <ThemedIcon iconType='MaterialIcons' name='arrow-right-alt' size={24} style={tailwind('px-1')} />
-          <ThemedText
-            light={tailwind('text-gray-500')}
-            dark={tailwind('text-gray-400')}
-            style={tailwind('text-sm')}
-            testID='convert_amount_target_suffix'
-          >
-            {targetUnit}
-          </ThemedText>
-        </SummaryTitle>
-      </ThemedView>
+          iconA='_UTXO'
+          fromAddress={address}
+          fromAddressLabel={addressLabel}
+        />
+        <NumberRowV2
+          containerStyle={{
+            style: tailwind('flex-row items-start w-full bg-transparent border-t-0.5 pt-5 mt-8'),
+            light: tailwind('bg-transparent border-mono-light-v2-300'),
+            dark: tailwind('bg-transparent border-mono-dark-v2-300')
+          }}
+          lhs={{
+            value: translate('screens/ConvertConfirmScreen', 'Transaction fee'),
+            testID: 'transaction_fee_label',
+            themedProps: {
+              light: tailwind('text-mono-light-v2-500'),
+              dark: tailwind('text-mono-dark-v2-500')
+            }
+          }}
+          rhs={{
+            value: fee.toFixed(8),
+            suffix: 'DFI',
+            testID: 'transaction_fee_value',
+            themedProps: {
+              light: tailwind('text-mono-light-v2-900'),
+              dark: tailwind('text-mono-dark-v2-900')
+            }
+          }}
+        />
 
-      <ThemedSectionTitle
-        testID='title_conversion_transaction_detail'
-        text={translate('screens/ConvertConfirmScreen', 'TRANSACTION DETAILS')}
-      />
-
-      <TextRow
-        lhs={translate('screens/ConvertConfirmScreen', 'Transaction type')}
-        rhs={{
-          value: translate('screens/ConvertConfirmScreen', 'Convert'),
-          testID: 'transaction_type'
-        }}
-        textStyle={tailwind('text-sm font-normal')}
-      />
-      <WalletAddressRow />
-      <NumberRow
-        lhs={translate('screens/ConvertConfirmScreen', '{{token}} to receive', { token: targetUnit })}
-        rhs={{
-          value: amount.toFixed(8),
-          testID: 'token_to_receive_amount'
-        }}
-      />
-
-      <InfoRow
-        type={InfoType.EstimatedFee}
-        value={fee.toFixed(8)}
-        testID='text_fee'
-        suffix='DFI'
-      />
-
-      <TransactionResultsRow
-        tokens={[
-          {
-            symbol: sourceUnit,
-            value: BigNumber.max(sourceBalance.minus(sourceUnit === 'UTXO' ? fee : 0), 0).toFixed(8)
+        <NumberRowV2
+          containerStyle={{
+            style: tailwind('flex-row items-start w-full bg-transparent mt-5'),
+            light: tailwind('bg-transparent'),
+            dark: tailwind('bg-transparent')
+          }}
+          lhs={{
+            value: translate('screens/ConvertConfirmScreen', 'Resulting Tokens'),
+            testID: 'resulting_tokens_label',
+            themedProps: {
+              light: tailwind('text-mono-light-v2-500'),
+              dark: tailwind('text-mono-dark-v2-500')
+            }
+          }} rhs={{
+          value: getResultingValue(ConvertTokenUnit.Token, fee, sourceBalance, sourceUnit, targetBalance, targetUnit),
+          suffix: 'DFI',
+          testID: 'resulting_tokens_value',
+          themedProps: {
+            light: tailwind('text-mono-light-v2-900 font-semibold-v2'),
+            dark: tailwind('text-mono-dark-v2-900 font-semibold-v2')
           },
-          {
-            symbol: targetUnit,
-            value: BigNumber.max(targetBalance.minus(targetUnit === 'UTXO' ? fee : 0), 0).toFixed(8)
+          subValue: {
+            value: getResultingPercentage(ConvertTokenUnit.Token, sourceBalance, sourceUnit, targetBalance),
+            prefix: '(',
+            suffix: '%)',
+            testID: 'resulting_tokens_sub_value'
           }
-        ]}
-      />
+        }}
+        />
 
-      <SubmitButtonGroup
-        isDisabled={isSubmitting || hasPendingJob || hasPendingBroadcastJob}
-        label={translate('screens/ConvertConfirmScreen', 'CONFIRM TRANSACTION')}
-        isProcessing={isSubmitting || hasPendingJob || hasPendingBroadcastJob}
-        processingLabel={translate('screens/ConvertConfirmScreen', 'CONVERTING')}
-        onCancel={onCancel}
-        onSubmit={onSubmit}
-        displayCancelBtn
-        title='convert'
-      />
-    </ThemedScrollView>
+        <NumberRowV2
+          containerStyle={{
+            style: tailwind('flex-row items-start w-full bg-transparent mt-5 border-b-0.5 pb-5'),
+            light: tailwind('bg-transparent border-mono-light-v2-300'),
+            dark: tailwind('bg-transparent border-mono-dark-v2-300')
+          }}
+          lhs={{
+            value: translate('screens/ConvertConfirmScreen', 'Resulting UTXO'),
+            testID: 'resulting_utxo_label',
+            themedProps: {
+              light: tailwind('text-mono-light-v2-500'),
+              dark: tailwind('text-mono-dark-v2-500')
+            }
+          }} rhs={{
+          value: getResultingValue(ConvertTokenUnit.UTXO, fee, sourceBalance, sourceUnit, targetBalance, targetUnit),
+          suffix: 'DFI',
+          testID: 'resulting_utxo_value',
+          themedProps: {
+            light: tailwind('text-mono-light-v2-900 font-semibold-v2'),
+            dark: tailwind('text-mono-dark-v2-900 font-semibold-v2')
+          },
+          subValue: {
+            value: getResultingPercentage(ConvertTokenUnit.UTXO, sourceBalance, sourceUnit, targetBalance),
+            prefix: '(',
+            suffix: '%)',
+            testID: 'resulting_utxo_sub_value'
+          }
+        }}
+        />
+
+        <View style={tailwind('mt-20')}>
+          <SubmitButtonGroupV2
+            isDisabled={false}
+            title='convert'
+            label={translate('screens/ConvertConfirmScreen', 'Convert')}
+            displayCancelBtn
+            onSubmit={onSubmit}
+            onCancel={onCancel}
+          />
+        </View>
+      </ThemedViewV2>
+    </ThemedScrollViewV2>
   )
 }
 
@@ -175,4 +198,18 @@ async function constructSignedConversionAndSend ({
   } catch (e) {
     logger.error(e)
   }
+}
+
+function getResultingValue (desireUnit: string, fee: BigNumber, balanceA: BigNumber, unitA: string, balanceB: BigNumber, unitB: string): string {
+  const balance = desireUnit === unitA ? balanceA : balanceB
+  const unit = desireUnit === unitA ? unitA : unitB
+
+  return BigNumber.max(balance.minus(unit === ConvertTokenUnit.UTXO ? fee : 0), 0).toFixed(8)
+}
+
+function getResultingPercentage (desireUnit: string, balanceA: BigNumber, unitA: string, balanceB: BigNumber): string {
+  const amount = desireUnit === unitA ? balanceA : balanceB
+  const totalAmount = balanceA.plus(balanceB)
+
+  return new BigNumber(amount).div(totalAmount).multipliedBy(100).toFixed(2)
 }
